move utility methods to ScalaReflection
